# open_api_docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/review-questions/{reviewQuestionInstance}/toggle-approval:
    post:
      tags:
      - rest-predefined-question-controller
      operationId: toggleApproval
      parameters:
      - name: reviewQuestionInstance
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
...
  /api/review-questions/{predefinedQuestion}/suggest-fine-tuning:
    post:
      tags:
      - rest-predefined-question-controller
      operationId: suggestQuestionForFineTuning
      parameters:
      - name: predefinedQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSuggestionCreationParams"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/review-questions/{note}/refine-question:
    post:
      tags:
      - rest-predefined-question-controller
      operationId: refineQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredefinedQuestion"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
  /api/review-questions/{note}/note-questions:
    get:
      tags:
      - rest-predefined-question-controller
      operationId: getAllQuestionByNote
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredefinedQuestion"
    post:
      tags:
      - rest-predefined-question-controller
      operationId: addQuestionManually
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredefinedQuestion"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
...
  /api/review-questions/generate-question-without-save:
    post:
      tags:
      - rest-predefined-question-controller
      operationId: generateQuestionWithoutSave
      parameters:
      - name: note
        in: query
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
...
  /api/review-questions/{note}/note-questions/{predefinedQuestion}:
    delete:
      tags:
      - rest-predefined-question-controller
      operationId: removeQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      - name: predefinedQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
    patch:
      tags:
      - rest-predefined-question-controller
      operationId: updateQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      - name: predefinedQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredefinedQuestion"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
...
components:
  schemas:
...
    MCQWithAnswer:
      required:
      - correctChoiceIndex
      - multipleChoicesQuestion
      type: object
      properties:
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctChoiceIndex:
          type: integer
          format: int32
    MultipleChoicesQuestion:
      required:
      - choices
      type: object
      properties:
        stem:
          type: string
        choices:
          type: array
          items:
            type: string
...
    BareQuestion:
      required:
      - multipleChoicesQuestion
      type: object
      properties:
        checkSpell:
          type: boolean
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
    PredefinedQuestion:
      required:
      - bareQuestion
      - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        bareQuestion:
          $ref: "#/components/schemas/BareQuestion"
        correctAnswerIndex:
          type: integer
          format: int32
        approved:
          type: boolean
...
    QuestionSuggestionCreationParams:
      type: object
      properties:
        comment:
          type: string
        isPositiveFeedback:
          type: boolean
    SuggestedQuestionForFineTuning:
      required:
      - id
      - preservedQuestion
      type: object
      properties:
        id:
          type: integer
          format: int32
        comment:
          type: string
        preservedQuestion:
          $ref: "#/components/schemas/MCQWithAnswer"
        preservedNoteContent:
          type: string
        realCorrectAnswers:
          type: string
        createdAt:
          type: string
          format: date-time
        positiveFeedback:
          type: boolean
...